backend/requirements.txt
django
django-cors-headers
djangorestframework
dotenv
drf-yasg
requests
Authlib